+++ tests/manual-test-examples/p5ar/basic/index.html
@@ -8,8 +8,6 @@
 
     <title>p5.XR AR Basic Example</title>
 
-    <script src='../../../../src/webxr/webxr-button.js'></script>
-
     <script src='../../../../node_modules/p5/lib/p5.min.js'></script>
     <script src='../../../../dist/p5xr.js'></script>
   </head>
