+++ tests/manual-test-examples/p5vr/basic/index.html
@@ -7,9 +7,6 @@
     <meta name='apple-mobile-web-app-capable' content='yes'>
 
     <title>p5.XR VR Basic Example</title>
-
-    <script src='../../../../src/webxr/webxr-button.js'></script>
-
     <script src='../../../../node_modules/p5/lib/p5.min.js'></script>
     <script src='../../../../dist/p5xr.js'></script>
   </head>
