+++ tests/manual-test-examples/p5vr/material/360PhotoViewer/index.html
@@ -8,7 +8,6 @@
     <title>p5.XR VR Custom Shader example</title>
 
     <link href='../../assets/common.css' rel='stylesheet'></link>
-    <script src='../../../../../src/webxr/webxr-button.js'></script>
     <script src='../../../../../node_modules/p5/lib/p5.min.js'></script>
     <script src='../../../../../dist/p5xr.js'></script>
   </head>
