+++ tests/manual-test-examples/p5vr/performance/spheres/index.html
@@ -7,7 +7,6 @@
     <meta name='apple-mobile-web-app-capable' content='yes'>
     <title>p5.XR Performance Testing</title>
 
-    <script src='../../../../../src/webxr/webxr-button.js'></script>
     <script src='../../../../../node_modules/p5/lib/p5.min.js'></script>
     <script src='../../../../../dist/p5xr.js'></script>
   </head>
