+++ tests/manual-test-examples/p5vr/raycasting/generateRay/index.html
@@ -8,8 +8,6 @@
 
     <title>generateRay raycasting example</title>
 
-    <script src='../../../../../src/webxr/webxr-button.js'></script>
-
     <script src='../../../../../node_modules/p5/lib/p5.js'></script>
     <script src='../../../../../dist/p5xr.js'></script>
   </head>
