+++ tests/manual-test-examples/p5vr/raycasting/intersectsPlane/index.html
@@ -8,8 +8,6 @@
 
     <title>intersectsPlane Raycasting Example</title>
 
-    <script src='../../../../../src/webxr/webxr-button.js'></script>
-
     <script src='../../../../../node_modules/p5/lib/p5.js'></script>
     <script src='../../../../../dist/p5xr.js'></script>
   </head>
