+++ tests/manual-test-examples/p5vr/raycasting/intersectsSphere/index.html
@@ -7,9 +7,6 @@
     <meta name='apple-mobile-web-app-capable' content='yes'>
 
     <title>intersectsSphere Raycasting example</title>
-
-    <script src='../../../../../src/webxr/webxr-button.js'></script>
-
     <script src='../../../../../node_modules/p5/lib/p5.js'></script>
     <script src='../../../../../dist/p5xr.js'></script>
   </head>
